Update the type of applyRefactorings, and update haddock
